requirements.txt
streamlit
openai
azure-search-documents
azure-core
azure-cosmos
python-dotenv
uuid